requirements.txt
mkdocs==1.4.0
wheel
GitPython==3.1.27